+++ client/src/helpers/twosky.js
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
 import twosky from 'MainRoot/.twosky.json';
 
 export const {