+++ internal/home/tls.go
@@ -210,15 +210,26 @@ type tlsConfigStatus struct {
 
 // field ordering is important -- yaml fields will mirror ordering from here
 type tlsConfig struct {
-	tlsConfigStatus   `json:",inline"`
+	tlsConfigStatus      `json:",inline"`
+	tlsConfigSettingsExt `json:",inline"`
+}
+
+// tlsConfigSettingsExt is used to (un)marshal PrivateKeySaved to ensure that
+// clients don't send and receive previously saved private keys.
+type tlsConfigSettingsExt struct {
 	tlsConfigSettings `json:",inline"`
+	// If private key saved as a string, we set this flag to true
+	// and omit key from answer.
+	PrivateKeySaved bool `yaml:"-" json:"private_key_saved,inline"`
 }
 
 func (t *TLSMod) handleTLSStatus(w http.ResponseWriter, _ *http.Request) {
 	t.confLock.Lock()
 	data := tlsConfig{
-		tlsConfigSettings: t.conf,
-		tlsConfigStatus:   t.status,
+		tlsConfigSettingsExt: tlsConfigSettingsExt{
+			tlsConfigSettings: t.conf,
+		},
+		tlsConfigStatus: t.status,
 	}
 	t.confLock.Unlock()
 	marshalTLS(w, data)
@@ -231,19 +242,23 @@ func (t *TLSMod) handleTLSValidate(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if setts.PrivateKeySaved {
+		setts.PrivateKey = t.conf.PrivateKey
+	}
+
 	if !WebCheckPortAvailable(setts.PortHTTPS) {
 		httpError(w, http.StatusBadRequest, "port %d is not available, cannot enable HTTPS on it", setts.PortHTTPS)
 		return
 	}
 
 	status := tlsConfigStatus{}
-	if tlsLoadConfig(&setts, &status) {
+	if tlsLoadConfig(&setts.tlsConfigSettings, &status) {
 		status = validateCertificates(string(setts.CertificateChainData), string(setts.PrivateKeyData), setts.ServerName)
 	}
 
 	data := tlsConfig{
-		tlsConfigSettings: setts,
-		tlsConfigStatus:   status,
+		tlsConfigSettingsExt: setts,
+		tlsConfigStatus:      status,
 	}
 	marshalTLS(w, data)
 }
@@ -290,16 +305,20 @@ func (t *TLSMod) handleTLSConfigure(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if data.PrivateKeySaved {
+		data.PrivateKey = t.conf.PrivateKey
+	}
+
 	if !WebCheckPortAvailable(data.PortHTTPS) {
 		httpError(w, http.StatusBadRequest, "port %d is not available, cannot enable HTTPS on it", data.PortHTTPS)
 		return
 	}
 
 	status := tlsConfigStatus{}
-	if !tlsLoadConfig(&data, &status) {
+	if !tlsLoadConfig(&data.tlsConfigSettings, &status) {
 		data2 := tlsConfig{
-			tlsConfigSettings: data,
-			tlsConfigStatus:   t.status,
+			tlsConfigSettingsExt: data,
+			tlsConfigStatus:      t.status,
 		}
 		marshalTLS(w, data2)
 
@@ -308,7 +327,7 @@ func (t *TLSMod) handleTLSConfigure(w http.ResponseWriter, r *http.Request) {
 
 	status = validateCertificates(string(data.CertificateChainData), string(data.PrivateKeyData), data.ServerName)
 
-	restartHTTPS := t.setConfig(data, status)
+	restartHTTPS := t.setConfig(data.tlsConfigSettings, status)
 	t.setCertFileTime()
 	onConfigModified()
 
@@ -320,8 +339,8 @@ func (t *TLSMod) handleTLSConfigure(w http.ResponseWriter, r *http.Request) {
 	}
 
 	data2 := tlsConfig{
-		tlsConfigSettings: data,
-		tlsConfigStatus:   t.status,
+		tlsConfigSettingsExt: data,
+		tlsConfigStatus:      t.status,
 	}
 
 	marshalTLS(w, data2)
@@ -335,7 +354,7 @@ func (t *TLSMod) handleTLSConfigure(w http.ResponseWriter, r *http.Request) {
 	// goroutine due to the same reason.
 	if restartHTTPS {
 		go func() {
-			Context.web.TLSConfigChanged(context.Background(), data)
+			Context.web.TLSConfigChanged(context.Background(), data.tlsConfigSettings)
 		}()
 	}
 }
@@ -514,8 +533,8 @@ func parsePrivateKey(der []byte) (crypto.PrivateKey, string, error) {
 }
 
 // unmarshalTLS handles base64-encoded certificates transparently
-func unmarshalTLS(r *http.Request) (tlsConfigSettings, error) {
-	data := tlsConfigSettings{}
+func unmarshalTLS(r *http.Request) (tlsConfigSettingsExt, error) {
+	data := tlsConfigSettingsExt{}
 	err := json.NewDecoder(r.Body).Decode(&data)
 	if err != nil {
 		return data, fmt.Errorf("failed to parse new TLS config json: %w", err)
@@ -559,8 +578,8 @@ func marshalTLS(w http.ResponseWriter, data tlsConfig) {
 	}
 
 	if data.PrivateKey != "" {
-		encoded := base64.StdEncoding.EncodeToString([]byte(data.PrivateKey))
-		data.PrivateKey = encoded
+		data.PrivateKeySaved = true
+		data.PrivateKey = ""
 	}
 
 	err := json.NewEncoder(w).Encode(data)